Seeding setup

+++ adventure/pkg/model/adventure.go
@@ -13,6 +13,6 @@ type Adventure struct {
 	Diet                 int32   `json:"diet,omitempty"`
 	Diet_hydration_notes string  `json:"diet_hydration_notes,omitempty"`
 	Beta_notes           string  `json:"beta_notes,omitempty"`
-	Lat                  float32 `json:"lat,omitempty"`
-	Lon                  float32 `json:"lon,omitempty"`
+	Lat                  float64 `json:"lat,omitempty"`
+	Lon                  float64 `json:"lon,omitempty"`
 }
